requirements.txt
fastapi>=0.115.0
google-auth>=2.35.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
requests>=2.31.0
uvicorn>=0.30.0
